requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
crispy-bootstrap5==0.6
Django==4.0.6
django-crispy-forms==1.14.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
lxml==4.9.1
Pillow==9.2.0
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
